requirements.txt
Flask==2.3.2
PyYAML>=6.0,<7.0
Werkzeug>=2.3.3